remove grpc dependency (ray_client_server) from conftest which is supposed to work with minimal
