+++ python/ray/tests/conftest.py
@@ -20,7 +20,6 @@ import pytest
 
 import ray
 import ray._private.ray_constants as ray_constants
-import ray.util.client.server.server as ray_client_server
 from ray._private.conftest_utils import set_override_dashboard_url  # noqa: F401
 from ray._private.runtime_env.pip import PipProcessor
 from ray._private.runtime_env.plugin_schema_manager import RuntimeEnvPluginSchemaManager
@@ -625,13 +624,6 @@ def call_ray_start_with_external_redis(request):
     # Delete the cluster address just in case.
     ray._private.utils.reset_ray_address()
 
-
-@pytest.fixture
-def init_and_serve():
-    server_handle, _ = ray_client_server.init_and_serve("localhost:50051")
-    yield server_handle
-    ray_client_server.shutdown_with_server(server_handle.grpc_server)
-    time.sleep(2)
 
 
 @pytest.fixture
